feat: add modules/cjsToEsm solution
