+++ src/modules/index.js
@@ -1,0 +1,5 @@
+import { createMyServer, unknownObject } from "./cjsToEsm.mjs";
+
+createMyServer.listen(() => {
+  console.log(`Unknown object: `, unknownObject);
+});
